refactor(orchestra): decompose server and dispatcher with label-driven dispatch
