+++ config/prompts.yaml
@@ -150,11 +150,18 @@ orchestra:
       {suggested_issue_details}
 
       ## 指令
-      请只围绕以下治理问题做判断：
-      1. 当前哪些 issue 已经在运行
-      2. 当前哪些 issue 值得建议启动
-      3. 是否需要最小 label 调整来表达上述判断
-      4. 如果暂不建议启动，说明原因
+      请通过修改 GitHub Issue 标签来表达治理决策。你的决策将由 Dispatcher 异步执行。
+
+      ### 决策输出规范（必须遵守）：
+      1. **推进 Issue**：如果一个 Issue 依赖已解除且准备好被执行，请在 GitHub Issue 上设置 `state/ready` 标签。
+      2. **标记阻塞**：如果一个 Issue 因为依赖或其他原因无法推进，请设置 `state/blocked` 标签，并发表评论说明原因。
+      3. **标记完成**：如果一个 Issue 对应的 PR 已合并或任务已确认完成，请设置 `state/done` 标签。
+      4. **无需操作**：对于已经在 `state/in-progress` 或 `state/review` 状态且运行正常的 Issue，保持现状。
+
+      **注意**：
+      - 不要直接修改 Assignee。
+      - 不要尝试直接调用执行命令或创建 Flow。
+      - 你的唯一输出手段是变更 GitHub Label。
 
   assignee_dispatch:
     # Template for the manager task instructions.
