+++ src/vibe3/orchestra/dispatcher.py
@@ -1,22 +1,29 @@
 """Command dispatcher with flow orchestration."""
 
 from pathlib import Path
+from typing import TYPE_CHECKING, Any
 
 from loguru import logger
 
 from vibe3.models.orchestration import IssueInfo, IssueState
 from vibe3.orchestra.circuit_breaker import CircuitBreaker
+from vibe3.orchestra.command_builder import CommandBuilder
 from vibe3.orchestra.config import OrchestraConfig
 from vibe3.orchestra.dispatcher_worktree import WorktreeResolverMixin
 from vibe3.orchestra.executor import run_command
 from vibe3.orchestra.flow_orchestrator import FlowOrchestrator
-from vibe3.orchestra.prompts import build_manager_command, render_manager_prompt
-from vibe3.prompts.models import PromptRecipe, PromptRenderResult
-from vibe3.prompts.template_loader import DEFAULT_PROMPTS_PATH
+from vibe3.orchestra.result_handler import DispatchResultHandler
+
+if TYPE_CHECKING:
+    from vibe3.prompts.models import PromptRecipe, PromptRenderResult
 
 
 class Dispatcher(WorktreeResolverMixin):
-    """Dispatches commands based on triggers with flow orchestration."""
+    """Dispatches commands based on triggers with flow orchestration.
+
+    Refactored to delegate command construction to CommandBuilder and
+    post-dispatch logic to DispatchResultHandler.
+    """
 
     def __init__(
         self,
@@ -31,7 +38,11 @@ class Dispatcher(WorktreeResolverMixin):
         self.dry_run = dry_run
         self.repo_path = repo_path or Path.cwd()
         self.orchestrator = orchestrator or FlowOrchestrator(config)
-        self._prompts_path = prompts_path or DEFAULT_PROMPTS_PATH
+
+        # Components
+        self.command_builder = CommandBuilder(config, prompts_path=prompts_path)
+        self.result_handler = DispatchResultHandler(config, self.orchestrator)
+
         # Initialize circuit breaker if enabled
         self._circuit_breaker = circuit_breaker
         if self._circuit_breaker is None and config.circuit_breaker.enabled:
@@ -40,16 +51,17 @@ class Dispatcher(WorktreeResolverMixin):
                 cooldown_seconds=config.circuit_breaker.cooldown_seconds,
                 half_open_max_tests=config.circuit_breaker.half_open_max_tests,
             )
+
         # Track last error category for feedback loop
         self._last_error_category: str | None = None
-        # Set after build_manager_command; used for dry-run provenance logging
-        self.last_manager_render_result: PromptRenderResult | None = None
+
+    @property
+    def last_manager_render_result(self) -> "PromptRenderResult | None":
+        """Proxy for CommandBuilder.last_manager_render_result for compatibility."""
+        return self.command_builder.last_manager_render_result
 
     def _run_command(self, cmd: list[str], cwd: Path, label: str) -> bool:
-        """Execute a subprocess command with timeout and structured logging.
-
-        Delegates to executor.run_command.
-        """
+        """Execute a subprocess command with timeout and structured logging."""
         success, category = run_command(
             cmd, cwd, label, circuit_breaker=self._circuit_breaker
         )
@@ -57,12 +69,7 @@ class Dispatcher(WorktreeResolverMixin):
         return success
 
     def run_governance_command(self, cmd: list[str], label: str) -> bool:
-        """Run a governance command through the shared dispatch machinery.
-
-        Uses the same circuit breaker and error classification as manager
-        dispatch, ensuring governance failures are tracked consistently.
-        Executes in repo_path (not a specific worktree).
-        """
+        """Run a governance command through the shared dispatch machinery."""
         return self._run_command(cmd, self.repo_path, label)
 
     @property
@@ -73,18 +80,14 @@ class Dispatcher(WorktreeResolverMixin):
         return "disabled"
 
     def dispatch_manager(self, issue: IssueInfo) -> bool:
-        """Dispatch manager execution for an assignee-triggered issue."""
+        """Dispatch manager execution for an issue."""
         log = logger.bind(
             domain="orchestra",
             action="manager_dispatch",
             issue=issue.number,
         )
 
-        render_result = render_manager_prompt(
-            self.config, issue, prompts_path=self._prompts_path
-        )
-        self.last_manager_render_result = render_result
-        cmd = build_manager_command(self.config, render_result.rendered_text)
+        cmd = self.command_builder.build_manager_command(issue)
 
         if self.dry_run:
             log.info("Dry run: skipping flow creation/label updates")
@@ -107,8 +110,7 @@ class Dispatcher(WorktreeResolverMixin):
             )
             return False
 
-        # Step 1.5: resolve manager execution cwd without mutating the
-        # serve process branch context.
+        # Step 1.5: resolve manager execution cwd
         manager_cwd = self._resolve_manager_cwd(issue.number, flow_branch)
         if not manager_cwd:
             log.error(
@@ -118,8 +120,8 @@ class Dispatcher(WorktreeResolverMixin):
             return False
         log.info(f"Manager dispatch using branch: {flow_branch} (cwd={manager_cwd})")
 
-        # Step 2: update display label
-        self._update_state_label(issue.number, IssueState.IN_PROGRESS)
+        # Step 2: update display label to in-progress (occupancy signal)
+        self.result_handler.update_state_label(issue.number, IssueState.IN_PROGRESS)
 
         # Step 3: run manager execution
         cmd = self._normalize_manager_command(cmd, manager_cwd)
@@ -129,21 +131,24 @@ class Dispatcher(WorktreeResolverMixin):
 
         # Step 4: feedback loop - update state based on result
         if success:
-            self._on_dispatch_success(issue, flow_branch)
+            self.result_handler.on_dispatch_success(issue, flow_branch)
         else:
-            self._on_dispatch_failure(issue, self._last_error_category or "unknown")
+            self.result_handler.on_dispatch_failure(
+                issue, self._last_error_category or "unknown"
+            )
 
         return success
 
     def dispatch_pr_review(self, pr_number: int) -> bool:
-        """Dispatch PR review command for reviewer-triggered events."""
+        """Dispatch PR review command."""
         log = logger.bind(
             domain="orchestra",
             action="review_dispatch",
             pr_number=pr_number,
         )
 
-        cmd, review_cwd = self.prepare_pr_review_dispatch(pr_number)
+        cmd = self.command_builder.build_pr_review_command(pr_number)
+        review_cwd = self._resolve_review_cwd_for_dispatch(pr_number)
         log.info(f"Dispatching review: {' '.join(cmd)} (cwd={review_cwd})")
 
         if self.dry_run:
@@ -152,206 +157,44 @@ class Dispatcher(WorktreeResolverMixin):
 
         return self._run_command(cmd, review_cwd, "Review execution")
 
+    def _resolve_review_cwd_for_dispatch(self, pr_number: int) -> Path:
+        """Helper to resolve PR review cwd based on config."""
+        if self.config.pr_review_dispatch.use_worktree:
+            return self.repo_path
+        return self._resolve_review_cwd(pr_number)
+
+    # --- Backward Compatibility Proxies for Tests ---
+
     def build_manager_command(self, issue: IssueInfo) -> list[str]:
-        """Proxy for prompts.build_manager_command, used in tests."""
-        render_result = render_manager_prompt(
-            self.config, issue, prompts_path=self._prompts_path
-        )
-        self.last_manager_render_result = render_result
-        return build_manager_command(self.config, render_result.rendered_text)
+        """Proxy for tests."""
+        return self.command_builder.build_manager_command(issue)
 
-    def _build_manager_recipe(self) -> PromptRecipe:
-        """Proxy for prompts.build_manager_recipe, used in tests."""
-        from vibe3.orchestra.prompts import build_manager_recipe
-
-        return build_manager_recipe(self.config)
+    def _build_manager_recipe(self) -> "PromptRecipe":
+        """Proxy for tests."""
+        return self.command_builder.build_manager_recipe()
 
     def prepare_pr_review_dispatch(self, pr_number: int) -> tuple[list[str], Path]:
-        """Prepare executable PR review command and working directory."""
-        cmd = ["uv", "run", "python", "-m", "vibe3", "review", "pr", str(pr_number)]
-        if self.config.pr_review_dispatch.async_mode:
-            cmd.append("--async")
-        if self.config.pr_review_dispatch.use_worktree:
-            cmd.append("--worktree")
-            review_cwd = self.repo_path
-        else:
-            review_cwd = self._resolve_review_cwd(pr_number)
-        return cmd, review_cwd
+        """Proxy for tests."""
+        cmd = self.command_builder.build_pr_review_command(pr_number)
+        cwd = self._resolve_review_cwd_for_dispatch(pr_number)
+        return cmd, cwd
 
     def _update_state_label(self, issue_number: int, state: IssueState) -> None:
-        """Update issue state label (display only, does not drive logic)."""
-        try:
-            from vibe3.services.label_service import LabelService
-
-            LabelService(repo=self.config.repo).confirm_issue_state(
-                issue_number,
-                state,
-                actor="orchestra:manager",
-                force=True,
-            )
-        except Exception as e:
-            logger.bind(domain="orchestra").warning(
-                f"Failed to update label for #{issue_number}: {e}"
-            )
+        """Proxy for tests."""
+        self.result_handler.update_state_label(issue_number, state)
 
     def _on_dispatch_success(self, issue: IssueInfo, flow_branch: str) -> None:
-        """Handle successful dispatch: check PR and update state to review.
-
-        Args:
-            issue: Issue that was dispatched
-            flow_branch: Flow branch name
-        """
-        log = logger.bind(
-            domain="orchestra",
-            action="dispatch_success",
-            issue=issue.number,
-        )
-
-        # Check if PR exists
-        pr_number = self.orchestrator.get_pr_for_issue(issue.number)
-        if pr_number:
-            # Update state to review
-            self._update_state_label(issue.number, IssueState.REVIEW)
-            log.info(
-                f"Dispatch success, PR #{pr_number} exists, "
-                f"advancing to state/review"
-            )
-
-            # Record event in flow history
-            self._record_dispatch_event(
-                flow_branch,
-                success=True,
-                issue_number=issue.number,
-                pr_number=pr_number,
-            )
-        else:
-            # No PR yet, keep in-progress state
-            log.info("Dispatch success, no PR yet, keeping state/in-progress")
-
-            # Record event in flow history
-            self._record_dispatch_event(
-                flow_branch,
-                success=True,
-                issue_number=issue.number,
-                pr_number=None,
-            )
+        """Proxy for tests."""
+        self.result_handler.on_dispatch_success(issue, flow_branch)
 
     def _on_dispatch_failure(self, issue: IssueInfo, category: str) -> None:
-        """Handle dispatch failure: update state and post comment.
-
-        Args:
-            issue: Issue that failed to dispatch
-            category: Error category (api_error, timeout, business_error, etc.)
-        """
-        log = logger.bind(
-            domain="orchestra",
-            action="dispatch_failure",
-            issue=issue.number,
-            category=category,
-        )
-
-        # For api_error and timeout, block the issue
-        if category in ("api_error", "timeout", "circuit_breaker"):
-            self._update_state_label(issue.number, IssueState.BLOCKED)
-            reason = f"Orchestra dispatch 失败（{category}），已暂停调度，等待恢复"
-            self._post_failure_comment(issue.number, reason)
-            log.warning(f"Issue blocked due to {category}")
-
-            # Record event in flow history
-            flow = self.orchestrator.get_flow_for_issue(issue.number)
-            if flow and flow.get("branch"):
-                self._record_dispatch_event(
-                    flow["branch"],
-                    success=False,
-                    issue_number=issue.number,
-                    category=category,
-                    reason=reason,
-                )
-            else:
-                log.warning(
-                    f"Cannot record dispatch event: no flow branch for #{issue.number}"
-                )
-        else:
-            # Business error - keep in-progress, don't auto-block
-            log.warning("Business error, keeping state/in-progress")
-
-            # Record event in flow history
-            flow = self.orchestrator.get_flow_for_issue(issue.number)
-            if flow and flow.get("branch"):
-                self._record_dispatch_event(
-                    flow["branch"],
-                    success=False,
-                    issue_number=issue.number,
-                    category=category,
-                    reason="Business logic error, manual intervention may be needed",
-                )
-            else:
-                log.warning(
-                    f"Cannot record dispatch event: no flow branch for #{issue.number}"
-                )
+        """Proxy for tests."""
+        self.result_handler.on_dispatch_failure(issue, category)
 
     def _post_failure_comment(self, issue_number: int, reason: str) -> None:
-        """Post failure comment on issue.
+        """Proxy for tests."""
+        self.result_handler.post_failure_comment(issue_number, reason)
 
-        Args:
-            issue_number: Issue number
-            reason: Failure reason
-        """
-        try:
-            from vibe3.clients.github_client import GitHubClient
-
-            GitHubClient().add_comment(
-                issue_number,
-                f"[Orchestra] {reason}",
-                repo=self.config.repo,
-            )
-        except Exception as exc:
-            logger.bind(domain="orchestra").warning(
-                f"Failed to post failure comment for #{issue_number}: {exc}"
-            )
-
-    def _record_dispatch_event(
-        self,
-        flow_branch: str,
-        success: bool,
-        issue_number: int,
-        pr_number: int | None = None,
-        category: str | None = None,
-        reason: str | None = None,
-    ) -> None:
-        """Record dispatch result as flow event for traceability.
-
-        Args:
-            flow_branch: Flow branch name
-            success: Whether dispatch succeeded
-            issue_number: Issue number
-            pr_number: PR number (if exists)
-            category: Error category (if failed)
-            reason: Failure reason (if failed)
-        """
-        try:
-            from vibe3.clients import SQLiteClient
-
-            store = SQLiteClient()
-            event_data: dict[str, int | str | None] = {
-                "success": success,
-                "issue": issue_number,
-            }
-            if pr_number is not None:
-                event_data["pr"] = pr_number
-            if category is not None:
-                event_data["category"] = category
-            if reason is not None:
-                event_data["reason"] = reason
-
-            store.add_event(
-                branch=flow_branch,
-                event_type="dispatch_result",
-                actor="orchestra:dispatcher",
-                detail="success" if success else f"failed:{category}",
-                refs=event_data,
-            )
-        except Exception as exc:
-            logger.bind(domain="orchestra").warning(
-                f"Failed to record dispatch event for #{issue_number}: {exc}"
-            )
+    def _record_dispatch_event(self, *args: Any, **kwargs: Any) -> None:
+        """Proxy for tests."""
+        self.result_handler.record_dispatch_event(*args, **kwargs)
