+++ src/vibe3/server/__init__.py
@@ -1,0 +1,1 @@
+"""vibe3 server module."""
