+++ tests/vibe3/orchestra/test_dispatcher_feedback.py
@@ -47,19 +47,17 @@ class TestDispatcherFeedbackLoop:
                         return_value=123,
                     ):
                         with patch.object(
-                            dispatcher, "_update_state_label"
+                            dispatcher.result_handler, "update_state_label"
                         ) as mock_update:
                             with patch.object(
-                                dispatcher, "_record_dispatch_event"
+                                dispatcher.result_handler, "record_dispatch_event"
                             ) as mock_record:
                                 result = dispatcher.dispatch_manager(issue)
 
         assert result is True
         # Should have called update_state_label with REVIEW (last call)
         review_calls = [
-            call
-            for call in mock_update.call_args_list
-            if call[0][1] == IssueState.REVIEW
+            c for c in mock_update.call_args_list if c[0][1] == IssueState.REVIEW
         ]
         assert len(review_calls) == 1
         # Should have recorded event
@@ -91,20 +89,20 @@ class TestDispatcherFeedbackLoop:
                         return_value=None,  # No PR
                     ):
                         with patch.object(
-                            dispatcher, "_update_state_label"
+                            dispatcher.result_handler, "update_state_label"
                         ) as mock_update:
-                            with patch.object(dispatcher, "_record_dispatch_event"):
+                            with patch.object(
+                                dispatcher.result_handler, "record_dispatch_event"
+                            ):
                                 result = dispatcher.dispatch_manager(issue)
 
         assert result is True
         # Should have called update_state_label with IN_PROGRESS (before execution)
         assert any(
-            call[0][1] == IssueState.IN_PROGRESS for call in mock_update.call_args_list
+            c[0][1] == IssueState.IN_PROGRESS for c in mock_update.call_args_list
         )
         # Should NOT have advanced to REVIEW
-        assert not any(
-            call[0][1] == IssueState.REVIEW for call in mock_update.call_args_list
-        )
+        assert not any(c[0][1] == IssueState.REVIEW for c in mock_update.call_args_list)
 
     def test_dispatch_failure_api_error_sets_blocked_and_comments(self):
         """API error should set issue to blocked and post comment."""
@@ -128,24 +126,26 @@ class TestDispatcherFeedbackLoop:
                         returncode=1, stderr="Error: rate limit exceeded"
                     ),
                 ):
-                    with patch.object(dispatcher, "_update_state_label") as mock_update:
+                    with patch.object(
+                        dispatcher.result_handler, "update_state_label"
+                    ) as mock_update:
                         with patch.object(
-                            dispatcher, "_post_failure_comment"
+                            dispatcher.result_handler, "post_failure_comment"
                         ) as mock_comment:
                             with patch.object(
                                 dispatcher.orchestrator,
                                 "get_flow_for_issue",
                                 return_value={"branch": "task/issue-102"},
                             ):
-                                with patch.object(dispatcher, "_record_dispatch_event"):
+                                with patch.object(
+                                    dispatcher.result_handler, "record_dispatch_event"
+                                ):
                                     result = dispatcher.dispatch_manager(issue)
 
         assert result is False
         # Should have set state to BLOCKED
         blocked_calls = [
-            call
-            for call in mock_update.call_args_list
-            if call[0][1] == IssueState.BLOCKED
+            c for c in mock_update.call_args_list if c[0][1] == IssueState.BLOCKED
         ]
         assert len(blocked_calls) >= 1
         # Should have posted comment
@@ -174,22 +174,26 @@ class TestDispatcherFeedbackLoop:
                     "subprocess.run",
                     side_effect=subprocess.TimeoutExpired(cmd=[], timeout=3600),
                 ):
-                    with patch.object(dispatcher, "_update_state_label") as mock_update:
-                        with patch.object(dispatcher, "_post_failure_comment"):
+                    with patch.object(
+                        dispatcher.result_handler, "update_state_label"
+                    ) as mock_update:
+                        with patch.object(
+                            dispatcher.result_handler, "post_failure_comment"
+                        ):
                             with patch.object(
                                 dispatcher.orchestrator,
                                 "get_flow_for_issue",
                                 return_value={"branch": "task/issue-103"},
                             ):
-                                with patch.object(dispatcher, "_record_dispatch_event"):
+                                with patch.object(
+                                    dispatcher.result_handler, "record_dispatch_event"
+                                ):
                                     result = dispatcher.dispatch_manager(issue)
 
         assert result is False
         # Should have set state to BLOCKED
         blocked_calls = [
-            call
-            for call in mock_update.call_args_list
-            if call[0][1] == IssueState.BLOCKED
+            c for c in mock_update.call_args_list if c[0][1] == IssueState.BLOCKED
         ]
         assert len(blocked_calls) >= 1
 
@@ -215,16 +219,20 @@ class TestDispatcherFeedbackLoop:
                         returncode=1, stderr="Error: merge conflict in file.py"
                     ),
                 ):
-                    with patch.object(dispatcher, "_update_state_label") as mock_update:
+                    with patch.object(
+                        dispatcher.result_handler, "update_state_label"
+                    ) as mock_update:
                         with patch.object(
-                            dispatcher, "_post_failure_comment"
+                            dispatcher.result_handler, "post_failure_comment"
                         ) as mock_comment:
                             with patch.object(
                                 dispatcher.orchestrator,
                                 "get_flow_for_issue",
                                 return_value={"branch": "task/issue-104"},
                             ):
-                                with patch.object(dispatcher, "_record_dispatch_event"):
+                                with patch.object(
+                                    dispatcher.result_handler, "record_dispatch_event"
+                                ):
                                     result = dispatcher.dispatch_manager(issue)
 
         assert result is False
@@ -232,9 +240,7 @@ class TestDispatcherFeedbackLoop:
         mock_comment.assert_not_called()
         # Should NOT have set state to BLOCKED
         blocked_calls = [
-            call
-            for call in mock_update.call_args_list
-            if call[0][1] == IssueState.BLOCKED
+            c for c in mock_update.call_args_list if c[0][1] == IssueState.BLOCKED
         ]
         assert len(blocked_calls) == 0
 
@@ -244,7 +250,7 @@ class TestDispatcherFeedbackLoop:
         dispatcher = Dispatcher(config, repo_path=Path("/tmp/repo"))
 
         # This test verifies the method exists and can be called
-        dispatcher._record_dispatch_event(
+        dispatcher.result_handler.record_dispatch_event(
             flow_branch="task/issue-100",
             success=True,
             issue_number=100,
@@ -252,7 +258,7 @@ class TestDispatcherFeedbackLoop:
         )
 
         # The method should exist and be callable
-        assert dispatcher._record_dispatch_event is not None
+        assert dispatcher.result_handler.record_dispatch_event is not None
 
     def test_post_failure_comment_creates_github_comment(self):
         """Failed dispatch should post comment on issue."""
@@ -265,7 +271,7 @@ class TestDispatcherFeedbackLoop:
             mock_github = MagicMock()
             mock_client_class.return_value = mock_github
 
-            dispatcher._post_failure_comment(100, "Test failure reason")
+            dispatcher.result_handler.post_failure_comment(100, "Test failure reason")
 
             # Verify the method was called (not checking internal implementation)
-            assert dispatcher._post_failure_comment is not None
+            assert dispatcher.result_handler.post_failure_comment is not None
